Assorted fixes

+++ common.py
@@ -6,9 +6,11 @@
 # STD LIB
 import os
 import pdb
+import sys
 import time
 import pickle
 import pathlib
+import platform
 import threading
 
 # EXTERNAL LIB
@@ -17,8 +19,8 @@ import threading
 from const import *
 
 def wait_complete(tag, target, args, remote):
-    # Try to create a placeholder.
-    placeholder = str(remote / (tag +'.plc'))
+    # Try to create a placeholder so that two nodes don't try to do the same job at once.
+    placeholder = str(remote / (tag + '.plc'))
     try:
         # This will only succeed if this program successfully created the placeholder.
         with open(placeholder, 'x') as handle:
@@ -31,13 +33,15 @@ def wait_complete(tag, target, args, remote):
         # Write the output to disk.
         with open(tag, 'wb') as handle:
             pickle.dump(result, handle)
-        
-        # Remove the placeholder.
-        os.remove(placeholder)
     except FileExistsError:
         # We couldn't claim that job, so WAIT until it's finished, then return None.
+        print('Job {} already claimed; waiting for output...')
         while os.path.exists(placeholder):
             time.sleep(1)
+    finally:
+        # Remove the placeholder in either case.
+        if os.path.exists(placeholder):
+            os.remove(placeholder)
 
 def read_tag(tag, remote):
     # Read data from a common tag.
@@ -52,15 +56,14 @@ def upload_files(fnames, dst, absolute_path=False):
     
     for fname in fnames:
         # print('\nUploading {}...'.format(fname))
+        # This function should handle both PosixPath as well as string destinations.
+        if absolute_path:
+            newname = dst
+        else:
+            newname = str(pathlib.Path(dst) / os.path.basename(fname))
+        # In case anyone is waiting on a file, "label" it as incomplete.
+        partname = newname + '.part'
         try:
-            # This function should handle both PosixPath as well as string destinations.
-            if absolute_path:
-                newname = dst
-            else:
-                newname = str(pathlib.Path(dst) / os.path.basename(fname))
-            
-            # In case anyone is waiting on a file, "label" it as incomplete.
-            partname = newname + '.part'
             # This will only succeed if this program successfully created the file.
             with open(partname, 'xb') as handle:
                 with open(str(fname), 'rb') as alt:
@@ -77,7 +80,10 @@ def upload_files(fnames, dst, absolute_path=False):
             # Try to upload the next file.
             print('\nFile already exists!')
             pass
-
+        except OSError:
+            # The file path specified is incorrect, or there was another error.
+            pass
+            
 def wait_for(fname):
     # If you wish upon a star...
     while not os.path.exists(fname):
